+++ schemas/request/investment.py
@@ -1,5 +1,6 @@
 import datetime
 import uuid
+from decimal import Decimal
 
 from fastapi import Query
 from pydantic import BaseModel, Field, model_validator
@@ -13,16 +14,16 @@ class CreateInvestmentRequest(BaseModel):
     type_id: uuid.UUID = Field(..., alias='typeId', description='The id of the investment type')
     transaction_date: datetime.date = Field(..., alias='transactionDate', description='The date of the investment')
     maturity_date: datetime.date = Field(None, alias='maturityDate', description='The date that the investment will be liquidated')
-    quantity: float = Field(None, description='The quantity of the investment bought')
-    price: float = Field(None, description='The unit price for the investment')
-    amount: float = Field(None, description='The total bought. Quantity * price')
+    quantity: Decimal = Field(None, description='The quantity of the investment bought')
+    price: Decimal = Field(None, description='The unit price for the investment')
+    amount: Decimal = Field(None, description='The total bought. Quantity * price')
     currency_id: str = Field(..., alias='currencyId', description='The id of the currency')
     indexer_type_id: uuid.UUID = Field(..., alias='indexerTypeId', description='The type of the index for the investment')
 
     indexer_id: uuid.UUID = Field(..., alias='indexerId', description='The id of the investment index')
     liquidity_id: uuid.UUID = Field(..., alias='liquidityId', description='The id of investment liquidity')
     liquidation_date: datetime.date = Field(None, alias='liquidationDate', description='The date that the investment was liquidated')
-    liquidation_amount: float = Field(None, alias='liquidationAmount', description='The amount liquidated, after tax')
+    liquidation_amount: Decimal = Field(None, alias='liquidationAmount', description='The amount liquidated, after tax')
     country_id: str = Field(..., alias='countryId', description='The id of the country')
 
     @model_validator(mode='before')
@@ -44,12 +45,12 @@ class GetInvestmentRequest(BaseModel):
 class LiquidateInvestmentRequest(BaseModel):
     id: uuid.UUID = Field(..., alias='investmentId', description='The unique identifier of the investment')
     liquidation_date: datetime.date = Field(None, alias='liquidationDate', description='The date that the investment was liquidated')
-    liquidation_amount: float = Field(None, alias='liquidationAmount', description='The amount liquidated, after tax')
+    liquidation_amount: Decimal = Field(None, alias='liquidationAmount', description='The amount liquidated, after tax')
 
 
 class TaxFeeRequest(BaseModel):
     id: uuid.UUID = Field(..., alias='taxFeeId', description='The identification of the tax/fee') # TODO: should be UUID not str
-    amount: float = Field(..., description='The amount of the tax/fee')
+    amount: Decimal = Field(..., description='The amount of the tax/fee')
     currency_id: str = Field('BRL', alias='currencyId', description='The currency of the tax/fee')
 
 
@@ -57,8 +58,8 @@ class CreateStatementRequest(BaseModel):
     investment_id: uuid.UUID = Field(..., alias='investmentId', description='The unique identifier of the investment')
     period: int = Field(..., alias='period', description='The period of the statement', examples=['202408'])
     reference_date: datetime.date = Field(..., alias='referenceDate', description='The date when the statement was calculated, usually the last business of the month')
-    gross_amount: float = Field(..., alias='grossAmount', description='The gross amount of the period')
-    net_amount: float = Field(..., alias='netAmount', description='The net amount of the period')
+    gross_amount: Decimal = Field(..., alias='grossAmount', description='The gross amount of the period')
+    net_amount: Decimal = Field(..., alias='netAmount', description='The net amount of the period')
     tax_detail: list[TaxFeeRequest] | None = Field(None, alias='taxDetail', description='The tax details of the investment tax')
     fee_detail: list[TaxFeeRequest] | None = Field(None, alias='feeDetail', description='The fee details of the investment fee')
 
@@ -70,7 +71,7 @@ class GetStatementRequest(BaseModel):
 class CreateObjectiveRequest(BaseModel):
     title: str = Field(..., alias='title', description='The title of the objective')
     description: str = Field(None, alias='description', description='The description of the objective')
-    amount: float = Field(None, alias='amount', description='The amount of the objective')
+    amount: Decimal = Field(None, alias='amount', description='The amount of the objective')
     estimated_deadline: datetime.date | None = Field(None, alias='estimatedDeadline', description='The estimated deadline of the objective')
